feat: loosen restrictions around semantic tags
This change combines the "semantic" and "non-semantic" tag workflows
into a single one which supports a wider range of "version tag"
specifications (that is, tags that match the regex `^v?\d+(\.\d+){0,2}[^
]*$`). A key behavior change that this introduces is support for action
repositories which do not use a fully qualified semantic version for
release tags, such as "v5", with no alternate like "v5.x.y".

The behavior of `parseTagLine` imitates the previous workflow, such that
we attempt to replace "non-semantic" version tags (but ultimately just
select the tag that matches the sha, regardless of its semantic
validity). Because these are "version strings" (like "v1" or "v1.2.3"),
the semantic conversion performed for comparison still works -- "v1"
would be seen as `1.0.0`, and when compared against `1.2.3`, the latter
would win. This results in a biased selection process which prefers
3-octect version strings over those that are not, without rejecting
them.

Closes: #939
Change-Id: Ibb13c4216a0682355a7d38880e607b6cb7d7ff03